@@ -1,65 +1,64 @@
 {
-  "name": "contributte/forms",
-  "description": "Extra contrib to nette/forms",
-  "keywords": [
-    "nette",
-    "forms",
-    "inputs",
-    "checkbox",
-    "rules",
-    "validation",
-    "renderers"
-  ],
-  "type": "library",
-  "license": "MIT",
-  "homepage": "https://github.com/contributte/forms",
-  "authors": [
-    {
-      "name": "Milan Felix Šulc",
-      "homepage": "https://f3l1x.io"
-    }
-  ],
-  "require": {
-    "php": ">=7.2",
-    "nette/forms": "^3.0.0"
-  },
-  "require-dev": {
-    "nette/application": "^3.0.0",
-    "nette/di": "^3.1.0",
-    "ninjify/nunjuck": "^0.4",
-    "ninjify/qa": "^0.12",
-    "phpstan/phpstan": "^1.0",
-    "phpstan/phpstan-deprecation-rules": "^1.0",
-    "phpstan/phpstan-nette": "^1.0",
-    "phpstan/phpstan-strict-rules": "^1.1"
-  },
-  "suggest": {
-    "nette/di": "to use FormFactoryExtension[CompilerExtension]"
-  },
-  "conflict": {
-    "nette/di": "<3.0.0"
-  },
-  "autoload": {
-    "psr-4": {
-      "Contributte\\Forms\\": "src"
-    }
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "Tests\\Fixtures\\": "tests/fixtures"
-    }
-  },
-  "config": {
-    "sort-packages": true,
-    "allow-plugins": {
-      "dealerdirect/phpcodesniffer-composer-installer": true
-    }
-  },
-  "minimum-stability": "dev",
-  "prefer-stable": true,
-  "extra": {
-    "branch-alias": {
-      "dev-master": "0.6.x-dev"
-    }
-  }
+	"name": "contributte/forms",
+	"description": "Extra contrib to nette/forms",
+	"keywords": [
+		"nette",
+		"forms",
+		"inputs",
+		"checkbox",
+		"rules",
+		"validation",
+		"renderers"
+	],
+	"type": "library",
+	"license": "MIT",
+	"homepage": "https://github.com/contributte/forms",
+	"authors": [
+		{
+			"name": "Milan Felix Šulc",
+			"homepage": "https://f3l1x.io"
+		}
+	],
+	"require": {
+		"php": ">=8.2",
+		"nette/forms": "^3.2"
+	},
+	"require-dev": {
+		"contributte/qa": "^0.4",
+		"contributte/phpstan": "^0.1",
+		"contributte/tester": "^0.3",
+		"mockery/mockery": "^1.6",
+		"nette/application": "^3.2",
+		"nette/di": "^3.2",
+		"nette/tester": "^2.5"
+	},
+	"suggest": {
+		"nette/di": "to use FormFactoryExtension[CompilerExtension]"
+	},
+	"conflict": {
+		"nette/di": "<3.2"
+	},
+	"autoload": {
+		"psr-4": {
+			"Contributte\\Forms\\": "src"
+		}
+	},
+	"autoload-dev": {
+		"psr-4": {
+			"Tests\\Fixtures\\": "tests/Fixtures"
+		}
+	},
+	"config": {
+		"sort-packages": true,
+		"allow-plugins": {
+			"dealerdirect/phpcodesniffer-composer-installer": true
+		}
+	},
+	"minimum-stability": "dev",
+	"prefer-stable": true,
+	"extra": {
+		"branch-alias": {
+			"dev-master": "0.7.x-dev"
+		}
+	}
 }
