@@ -11,39 +11,34 @@ use Nette\Utils\Html;
 class DateTimeInput extends AbstractDateTimeInput
 {
 
-	protected $defaultHumanFormat = 'Y-m-d H:i';
+	protected string $defaultHumanFormat = 'Y-m-d H:i';
 
-	protected $htmlFormat = 'Y-m-d\TH:i';
+	protected string $htmlFormat = 'Y-m-d\TH:i';
 
-	protected $htmlClass = 'datetime-input';
+	protected string $htmlClass = 'datetime-input';
 
-	/** @var ?DateTimeZone */
-	protected $inputTimezone = null;
+	protected ?DateTimeZone $inputTimezone = null;
 
 	/**
 	 * @param string|string[] $format
 	 */
-	public function __construct(?string $label = null, $format = null, ?DateTimeZone $inputTimezone = null)
+	public function __construct(?string $label = null, string|array|null $format = null, ?DateTimeZone $inputTimezone = null)
 	{
 		parent::__construct($label, $format);
+
 		$this->parser->setDefaultTimezone($inputTimezone);
 		$this->inputTimezone = $inputTimezone;
 	}
 
-	protected function normalizeDateTime(DateTimeInterface $datetime): DateTimeImmutable
+	public static function register(?string $defaultFormat = null): void
 	{
-		$datetime = parent::normalizeDateTime($datetime);
-		if ($this->inputTimezone !== null) {
-			$datetime = $datetime->setTimezone($this->inputTimezone);
-		}
-
-		return $datetime;
+		Container::extensionMethod(
+			'addDateTime',
+			fn ($container, $name, $label = null, $format = null, ?DateTimeZone $inputTimezone = null): DateTimeInput => $container[$name] = new DateTimeInput($label, $format ?? $defaultFormat, $inputTimezone)
+		);
 	}
 
-	/**
-	 * @return mixed
-	 */
-	public function getValueInTz(?DateTimeZone $timezone = null)
+	public function getValueInTz(?DateTimeZone $timezone = null): mixed
 	{
 		$value = $this->getValueAsDateTimeImmutable();
 		if ($value === null) {
@@ -55,17 +50,16 @@ class DateTimeInput extends AbstractDateTimeInput
 		}
 
 		$value = $value->setTimezone($timezone);
+
 		return $this->transformValue($value);
 	}
 
 	/**
 	 * @template T
-	 * @param string|callable $type
 	 * @phpstan-param class-string<T>|callable(DateTimeInterface): ?T $type
-	 * @return mixed
 	 * @phpstan-return ?T
 	 */
-	public function getValueInTzAs($type, ?DateTimeZone $timezone = null)
+	public function getValueInTzAs(string|callable $type, ?DateTimeZone $timezone = null): mixed
 	{
 		$value = $this->getValueAsDateTimeImmutable();
 		if ($value === null) {
@@ -100,14 +94,14 @@ class DateTimeInput extends AbstractDateTimeInput
 		return $control;
 	}
 
-	public static function register(?string $defaultFormat = null): void
+	protected function normalizeDateTime(DateTimeInterface $datetime): DateTimeImmutable
 	{
-		Container::extensionMethod(
-			'addDateTime',
-			function ($container, $name, $label = null, $format = null, ?DateTimeZone $inputTimezone = null) use ($defaultFormat): DateTimeInput {
-				return $container[$name] = new DateTimeInput($label, $format ?? $defaultFormat, $inputTimezone);
-			}
-		);
+		$datetime = parent::normalizeDateTime($datetime);
+		if ($this->inputTimezone !== null) {
+			$datetime = $datetime->setTimezone($this->inputTimezone);
+		}
+
+		return $datetime;
 	}
 
 }
