@@ -9,8 +9,7 @@ use Nette\Forms\Rendering\DefaultFormRenderer;
 class Bootstrap3InlineRenderer extends DefaultFormRenderer
 {
 
-	/** @var mixed[] */
-	public $wrappers = [
+	public array $wrappers = [
 		'form' => [
 			'container' => '',
 		],
@@ -64,7 +63,7 @@ class Bootstrap3InlineRenderer extends DefaultFormRenderer
 	 * @param string|null $mode 'begin', 'errors', 'ownerrors', 'body', 'end' or empty to render all
 	 * @phpcsSuppress SlevomatCodingStandard.TypeHints.TypeHintDeclaration.MissingParameterTypeHint
 	 */
-	public function render(Form $form, $mode = null): string
+	public function render(Form $form, ?string $mode = null): string
 	{
 		$form->getElementPrototype()->addClass('form-inline');
 
@@ -86,7 +85,7 @@ class Bootstrap3InlineRenderer extends DefaultFormRenderer
 				case $control instanceof Controls\Checkbox:
 				case $control instanceof Controls\CheckboxList:
 				case $control instanceof Controls\RadioList:
-					$control->getSeparatorPrototype()->setName('div')->addClass($control->getControlPrototype()->type);
+					$control->getContainerPrototype()->setName('div')->addClass($control->getControlPrototype()->type);
 					break;
 			}
 		}
