@@ -2,16 +2,15 @@
 
 namespace Contributte\Forms\Rendering;
 
+use Nette\Forms\Control;
 use Nette\Forms\Controls;
 use Nette\Forms\Form;
-use Nette\Forms\IControl;
 use Nette\Utils\Html;
 
 class Bootstrap4HorizontalRenderer extends AbstractBootstrapHorizontalRenderer
 {
 
-	/** @var mixed[] */
-	public $wrappers = [
+	public array $wrappers = [
 		'form' => [
 			'container' => null,
 		],
@@ -64,7 +63,7 @@ class Bootstrap4HorizontalRenderer extends AbstractBootstrapHorizontalRenderer
 	 * @param string|null $mode 'begin', 'errors', 'ownerrors', 'body', 'end' or empty to render all
 	 * @phpcsSuppress SlevomatCodingStandard.TypeHints.TypeHintDeclaration.MissingParameterTypeHint
 	 */
-	public function render(Form $form, $mode = null): string
+	public function render(Form $form, ?string $mode = null): string
 	{
 		$form->getElementPrototype()->setNovalidate(true);
 
@@ -93,7 +92,7 @@ class Bootstrap4HorizontalRenderer extends AbstractBootstrapHorizontalRenderer
 				case $control instanceof Controls\Checkbox:
 				case $control instanceof Controls\CheckboxList:
 				case $control instanceof Controls\RadioList:
-					$control->getSeparatorPrototype()->setName('div')->addClass('form-check');
+					$control->getContainerPrototype()->setName('div')->addClass('form-check');
 					$control->getControlPrototype()->addClass('form-check-input');
 					$control->getLabelPrototype()->addClass('form-check-label');
 					break;
@@ -103,7 +102,7 @@ class Bootstrap4HorizontalRenderer extends AbstractBootstrapHorizontalRenderer
 		return parent::render($form, $mode);
 	}
 
-	public function renderLabel(IControl $control): Html
+	public function renderLabel(Control $control): Html
 	{
 		$label = parent::renderLabel($control);
 		if ($control instanceof Controls\Checkbox || $control instanceof Controls\CheckboxList || $control instanceof Controls\RadioList) {
