@@ -7,18 +7,22 @@ use Nette\Forms\Form;
 final class TestForm extends Form
 {
 
+	/**
+	 * @param array<string, mixed> $values
+	 */
 	public function __construct(array $values)
 	{
 		parent::__construct();
+
 		$this->allowCrossOrigin();
 		$_SERVER['REQUEST_METHOD'] = 'POST';
 		$_POST = $values;
 		$_POST['btn'] = '';
-		$this->addSubmit('btn')->onClick[] = function () {
+		$this->addSubmit('btn')->onClick[] = function (): void {
 		};
 	}
 
-	public function submit()
+	public function submit(): void
 	{
 		Form::initialize(true);
 		$this->fireEvents();
