@@ -1,13 +1,17 @@
 <?php declare(strict_types = 1);
 
-use Ninjify\Nunjuck\Environment;
+use Contributte\Tester\Environment;
 
 if (@!include __DIR__ . '/../vendor/autoload.php') {
 	echo 'Install Nette Tester using `composer update --dev`';
 	exit(1);
 }
 
-// Configure environment
-Environment::setupTester();
-Environment::setupTimezone();
-Environment::setupVariables(__DIR__);
+Environment::setup(__DIR__);
+
+define('TEMP_DIR', Environment::getTestDir());
+
+function test(Closure $function): void
+{
+	$function();
+}
